Fix close state of ReadWriteZipFS
ReadWriteZipFS.isClosed() would always return false, because the
protected "close" field of the super class was never set to true.
This allows to enable the close check for LIBRARY_OUTPUT_ZIP.

+++ sched/src/com/android/sched/vfs/ReadWriteZipFS.java
@@ -90,14 +90,17 @@ public class ReadWriteZipFS extends BaseVFS<BaseVDir, BaseVFile> implements VFS 
   }
 
   @Override
-  public void close() throws CannotCloseException {
-    vfs.close();
-    try {
-      if (dir.exists()) {
-        FileUtils.deleteDir(dir);
+  public synchronized void close() throws CannotCloseException {
+    if (!closed) {
+      vfs.close();
+      try {
+        if (dir.exists()) {
+          FileUtils.deleteDir(dir);
+        }
+      } catch (IOException e) {
+        throw new CannotCloseException(new DirectoryLocation(dir), e);
       }
-    } catch (IOException e) {
-      throw new CannotCloseException(new DirectoryLocation(dir), e);
+      closed = true;
     }
   }
